Move constants from types file to own constants file
