+++ ui/src/reusable_ui/components/wizard/ProgressConnector.tsx
@@ -1,5 +1,5 @@
 import React, {PureComponent} from 'react'
-import {ConnectorState} from 'src/types/wizard'
+import {ConnectorState} from 'src/reusable_ui/constants/wizard'
 import {ErrorHandling} from 'src/shared/decorators/errors'
 
 import 'src/reusable_ui/components/wizard/ProgressConnector.scss'
