+++ ui/src/reusable_ui/components/wizard/WizardCloak.tsx
@@ -1,7 +1,8 @@
 import React, {PureComponent, ReactElement} from 'react'
 import WizardProgressBar from 'src/reusable_ui/components/wizard/WizardProgressBar'
 import {ErrorHandling} from 'src/shared/decorators/errors'
-import {WizardStepProps, StepStatus, Step} from 'src/types/wizard'
+import {WizardStepProps, Step} from 'src/types/wizard'
+import {StepStatus} from 'src/reusable_ui/constants/wizard'
 
 import 'src/reusable_ui/components/wizard/WizardCloak.scss'
