+++ ui/src/reusable_ui/components/wizard/test/ProgressConnector.test.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import {shallow} from 'enzyme'
 
 import ProgressConnector from 'src/reusable_ui/components/wizard/ProgressConnector'
-import {ConnectorState} from 'src/types/wizard'
+import {ConnectorState} from 'src/reusable_ui/constants/wizard'
 
 describe('Progress Connector', () => {
   let wrapper
