+++ ui/src/reusable_ui/constants/wizard.ts
@@ -1,0 +1,11 @@
+export enum ConnectorState {
+  None = 'none',
+  Some = 'some',
+  Full = 'full',
+}
+
+export enum StepStatus {
+  Incomplete = 'circle-thick',
+  Complete = 'checkmark',
+  Error = 'remove',
+}
